requirements.txt
tqdm
python-box
pillow
pathlib
pyyaml
mlflow
dvc
dvc-s3

--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision

